Revert "Revert "[1753] Replace custom routing with react-instantsearch-router-nextjs""
This reverts commit 8a851058ee829eaf95b630cbc5ddbb94348f98e5.
