+++ components/search/bills/BillSearch.tsx
@@ -7,21 +7,23 @@ import {
   SearchBox,
   useInstantSearch
 } from "react-instantsearch"
-import { currentGeneralCourt } from "functions/src/shared"
+import { createInstantSearchRouterNext } from "react-instantsearch-router-nextjs"
+import singletonRouter from "next/router"
 import styled from "styled-components"
 import TypesenseInstantSearchAdapter from "typesense-instantsearch-adapter"
+import { currentGeneralCourt } from "functions/src/shared"
 import { Col, Container, Row, Spinner } from "../../bootstrap"
 import { NoResults } from "../NoResults"
 import { ResultCount } from "../ResultCount"
 import { SearchContainer } from "../SearchContainer"
 import { SearchErrorBoundary } from "../SearchErrorBoundary"
-import { useRouting } from "../useRouting"
 import { BillHit } from "./BillHit"
 import { useBillRefinements } from "./useBillRefinements"
 import { SortBy, SortByWithConfigurationItem } from "../SortBy"
 import { getServerConfig, VirtualFilters } from "../common"
 import { useBillSort } from "./useBillSort"
 import { FC, useState } from "react"
+import { pathToSearchState, searchStateToUrl } from "../routingHelpers"
 
 const searchClient = new TypesenseInstantSearchAdapter({
   server: getServerConfig(),
@@ -70,8 +72,16 @@ export const BillSearch = () => {
           }
         }}
         searchClient={searchClient}
-        routing={useRouting()}
-        future={{ preserveSharedStateOnUnmount: true }}
+        routing={{
+          router: createInstantSearchRouterNext({
+            singletonRouter,
+            routerOptions: {
+              cleanUrlOnDispose: false,
+              createURL: args => searchStateToUrl(args),
+              parseURL: args => pathToSearchState(args)
+            }
+          })
+        }}
       >
         <VirtualFilters type="bill" />
         <Layout items={items} />
