+++ package.json
@@ -118,6 +118,7 @@
     "react-i18next": "^13.2.2",
     "react-inlinesvg": "^3.0.1",
     "react-instantsearch": "^7.12.4",
+    "react-instantsearch-router-nextjs": "^7.15.5",
     "react-is": "^18.2.0",
     "react-markdown": "^8.0.4",
     "react-overlays": "^5.1.1",
